app/requirements.txt
PyYAML
numpy
qiskit-optimization
psutil
matplotlib
qiskit==1.1.0